refactor: organize the code in different modules
We wants the code be more readable.
