+++ src/window/mod.rs
@@ -1,0 +1,5 @@
+pub mod rstk_ext;
+pub mod toolkit;
+pub mod tooltip;
+
+use super::config;
